pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.6.2",
    "django>=6.0.1",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.16.1",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "vt-py>=0.22.0",
    "daphne>=4.2.1",
    "urlscan-python>=0.0.2",
    "pillow>=12.1.0",
    "gevent>=25.9.1",
    "google-genai>=1.62.0",
]
